+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v1.3.0 (July 31st, 2018)
+- Added an optional `page` parameter to `Cards.all()`, defaulting to `1`. To get only the one page, you can call `cancelAfterPage()` on the resulting `MagicEmitter`.
+- Added an optional `lang` parameter to [`Cards.bySet()`](./README.md#cardsbyset-code-string-collectorid-number-lang-string-promisecard-). See the [Scryfall Documentation for a list of valid languages](https://scryfall.com/docs/api/languages). 
+- Added support for [`/cards/arena/:id`](./README.md#cardsbyarenaid-id-number-promisecard-)
+
 # v1.2.1 (June 14th, 2018)
 Fixed `waitForAll` throwing an uncatchable promise rejection error.